pyproject.toml
[project]
name = "stayawake"
version = "0.3"
description = ""
authors = [
    {name = "Brett Heinkel", email = "brett@heinkel.dev"},
]
dependencies = [
    "pillow>=10.0.1",
    "pystray>=0.19.5",
    "pyautogui>=0.9.54",
]
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "pyinstaller>=5.13.2",
    "flake8>=6.1.0",
]
